+++ packages/next/test/fixtures/00-app-dir-actions-experimental-streaming/next.config.js
@@ -9,6 +9,18 @@ module.exports = {
         source: '/rewrite/edge/rsc/static',
         destination: '/edge/rsc/static',
       },
+      {
+        source: '/greedy-rewrite/static/:path*',
+        destination: '/static/:path*',
+      },
+      {
+        source: '/greedy-rewrite/edge/static/:path*',
+        destination: '/edge/static/:path*',
+      },
+      {
+        source: '/rewritten-to-index',
+        destination: '/?fromRewrite=1',
+      },
     ];
   },
 };